+++ src/js/api.ts
@@ -8,7 +8,7 @@ import { version } from "./version";
 import { setStdin, setStdout, setStderr } from "./streams";
 import { scheduleCallback } from "./scheduler";
 import { TypedArray, PackageData, FSType, Lockfile } from "./types";
-import { IN_NODE, detectEnvironment } from "./environments";
+import { RUNTIME_ENV } from "./environments";
 // @ts-ignore
 import LiteralMap from "./common/literal-map";
 import abortSignalAny from "./common/abortSignalAny";
@@ -69,9 +69,6 @@ API.restoreState = (state: any) => API.pyodide_py._state.restore_state(state);
 /** @private */
 API.scheduleCallback = scheduleCallback;
 
-/** @private */
-API.detectEnvironment = detectEnvironment;
-
 // @ts-ignore
 if (typeof AbortSignal !== "undefined" && AbortSignal.any) {
   /** @private */
@@ -90,10 +87,10 @@ function ensureMountPathExists(path: string): void {
     follow_mount: false,
   });
 
-  if (FS.isMountpoint(node)) {
+  if (Module.FS.isMountpoint(node)) {
     throw new Error(`path '${path}' is already a file system mount point`);
   }
-  if (!FS.isDir(node.mode)) {
+  if (!Module.FS.isDir(node.mode)) {
     throw new Error(`path '${path}' points to a file not a directory`);
   }
   for (const _ in node.contents) {
@@ -579,7 +576,7 @@ export class PyodideAPI_ {
    * @param hostPath The host path to mount. It must be a directory that exists.
    */
   static mountNodeFS(emscriptenPath: string, hostPath: string): void {
-    if (!IN_NODE) {
+    if (!RUNTIME_ENV.IN_NODE) {
       throw new Error("mountNodeFS only works in Node");
     }
     ensureMountPathExists(emscriptenPath);