+++ src/js/types.ts
@@ -3,10 +3,10 @@ import type { PyProxy, PyAwaitable } from "generated/pyproxy";
 import { type PyodideAPI } from "./api";
 import { type ConfigType } from "./pyodide";
 import { type InFuncType } from "./streams";
+import { type RuntimeEnv } from "./environments";
 import { SnapshotConfig } from "./snapshot";
 import { ResolvablePromise } from "./common/resolveable";
 import { PackageManager } from "./load-package";
-
 /**
  * @docgroup pyodide.ffi
  */
@@ -22,7 +22,7 @@ export type TypedArray =
   | Float64Array;
 
 declare global {
-  export var Module: Module;
+  export var Module: PyodideModule;
   export var API: API;
 }
 
@@ -32,12 +32,11 @@ declare global {
 // not necessary to put them in `-s EXPORTED_RUNTIME_METHODS`.
 declare global {
   export const stringToNewUTF8: (str: string) => number;
-  export const UTF8ToString: (ptr: number) => string;
+  // export const UTF8ToString: (ptr: number) => string; => removed due to duplicate with @types/emscripten
   export const UTF8ArrayToString: (buf: Uint8Array) => string;
-  export const FS: FSType;
-  export const stackAlloc: (sz: number) => number;
-  export const stackSave: () => number;
-  export const stackRestore: (ptr: number) => void;
+  // export const stackAlloc: (sz: number) => number; => removed due to duplicate with @types/emscripten
+  // export const stackSave: () => number; => removed due to duplicate with @types/emscripten
+  // export const stackRestore: (ptr: number) => void; => removed due to duplicate with @types/emscripten
   export const HEAPU32: Uint32Array;
 }
 
@@ -226,13 +225,15 @@ export type FSStreamOpsGen<T> = {
 };
 
 /**
+ * Methods that the Emscripten filesystem provides. Most of them are already defined
+ * in `@types/emscripten`, but Pyodide uses quite a lot of private APIs that are not
+ * defined there as well. Hence this interface.
+ *
+ * TODO: Consider upstreaming these APIs to `@types/emscripten`.
  * @hidden
  */
-export interface FSType {
-  unlink: (path: string) => void;
+interface PyodideFSType {
   mkdirTree: (path: string, mode?: number) => void;
-  chdir: (path: string) => void;
-  symlink: (target: string, src: string) => FSNode;
   createDevice: ((
     parent: string,
     name: string,
@@ -241,43 +242,29 @@ export interface FSType {
   ) => FSNode) & {
     major: number;
   };
-  closeStream: (fd: number) => void;
-  open: (path: string, flags: string | number, mode?: number) => FSStream;
-  makedev: (major: number, minor: number) => number;
-  mkdev: (path: string, dev: number) => FSNode;
-  filesystems: any;
-  stat: (path: string, dontFollow?: boolean) => any;
-  readdir: (path: string) => string[];
-  isDir: (mode: number) => boolean;
-  isMountpoint: (mode: FSNode) => boolean;
   lookupPath: (
     path: string,
     options?: {
       follow_mount?: boolean;
     },
   ) => { node: FSNode };
-  isFile: (mode: number) => boolean;
-  writeFile: (path: string, contents: any, o?: { canOwn?: boolean }) => void;
-  chmod: (path: string, mode: number) => void;
-  utime: (path: string, atime: number, mtime: number) => void;
-  rmdir: (path: string) => void;
-  mount: (type: any, opts: any, mountpoint: string) => any;
-  unmount: (mountpoint: string) => any;
-  write: (
-    stream: FSStream,
-    buffer: any,
-    offset: number,
-    length: number,
-    position?: number,
-  ) => number;
-  close: (stream: FSStream) => void;
-  ErrnoError: { new (errno: number): Error };
+  open: (path: string, flags: string | number, mode?: number) => FSStream;
+  filesystems: any;
+  isMountpoint: (node: FSNode) => boolean;
+  closeStream: (fd: number) => void;
   registerDevice<T>(dev: number, ops: FSStreamOpsGen<T>): void;
-  syncfs(dir: boolean, oncomplete: (val: void) => void): void;
+  writeFile: (path: string, contents: any, o?: { canOwn?: boolean }) => void;
 }
 
+/**
+ * Combined filesystem type that omits the incompatible lookupPath from `@types/emscripten` and adds Pyodide-specific filesystem methods.
+ * TODO: Consider upstreaming these APIs to `@types/emscripten`
+ * @hidden
+ */
+export type FSType = Omit<typeof FS, "lookupPath"> & PyodideFSType;
+
 /** @hidden */
-export type PreRunFunc = (Module: Module) => void;
+export type PreRunFunc = (Module: PyodideModule) => void;
 
 type DSO = any;
 
@@ -288,13 +275,8 @@ export interface LDSO {
   };
 }
 
-/**
- * TODO: consider renaming the type to ModuleType to avoid name collisions
- * between Module and ModuleType?
- * @hidden
- */
-export interface Module {
-  API: API;
+/** @hidden */
+export interface EmscriptenModule {
   locateFile: (file: string) => string;
   exited?: { toThrow: any };
   ENV: { [key: string]: string };
@@ -312,6 +294,36 @@ export interface Module {
   ERRNO_CODES: { [k: string]: number };
   stringToNewUTF8(x: string): number;
   stringToUTF8OnStack: (str: string) => number;
+  HEAP8: Uint8Array;
+  HEAPU32: Uint32Array;
+  getExceptionMessage(e: number): [string, string];
+  exitCode: number | undefined;
+  ExitStatus: { new (exitCode: number): Error };
+  _free: (ptr: number) => void;
+  stackSave: () => number;
+  stackRestore: (ptr: number) => void;
+  promiseMap: {
+    free(id: number): void;
+  };
+  _emscripten_dlopen_promise(lib: number, flags: number): number;
+  _dlerror(): number;
+  UTF8ToString: (
+    ptr: number,
+    maxBytesToRead: number,
+    ignoreNul?: boolean,
+  ) => string;
+}
+
+/** @hidden */
+export interface PythonModule extends EmscriptenModule {
+  _Py_EMSCRIPTEN_SIGNAL_HANDLING: number;
+  Py_EmscriptenSignalBuffer: TypedArray;
+  _Py_Version: number;
+}
+
+/** @hidden */
+export interface PyodideModule extends PythonModule {
+  API: API;
   _compat_to_string_repr: number;
   _compat_null_to_none: number;
   js2python_convert: (
@@ -326,36 +338,16 @@ export interface Module {
     },
   ) => any;
   _PropagatePythonError: typeof Error;
-  _Py_EMSCRIPTEN_SIGNAL_HANDLING: number;
-  Py_EmscriptenSignalBuffer: TypedArray;
-  HEAP8: Uint8Array;
-  HEAPU32: Uint32Array;
   __hiwire_get(a: number): any;
   __hiwire_set(a: number, b: any): void;
   __hiwire_immortal_add(a: any): void;
   _jslib_init(): number;
   _init_pyodide_proxy(): number;
-  getExceptionMessage(e: number): [string, string];
+
   handle_js_error(e: any): void;
-  exitCode: number | undefined;
-  ExitStatus: { new (exitCode: number): Error };
-  _Py_Version: number;
   _print_stdout: (ptr: number) => void;
   _print_stderr: (ptr: number) => void;
-  _free: (ptr: number) => void;
-  stackSave: () => number;
-  stackRestore: (ptr: number) => void;
-  promiseMap: {
-    free(id: number): void;
-  };
-  _emscripten_dlopen_promise(lib: number, flags: number): number;
   getPromise(p: number): Promise<any>;
-  _dlerror(): number;
-  UTF8ToString: (
-    ptr: number,
-    maxBytesToRead: number,
-    ignoreNul?: boolean,
-  ) => string;
 }
 
 /**
@@ -477,6 +469,7 @@ export type PackageLoadMetadata = {
 
 /** @hidden */
 export interface API {
+  runtimeEnv: RuntimeEnv;
   fatal_error: (e: any) => never;
   isPyProxy: (e: any) => e is PyProxy;
   debug_ffi: boolean;
@@ -518,7 +511,6 @@ export interface API {
   saveState: () => any;
   restoreState: (state: any) => void;
   scheduleCallback: (callback: () => void, timeout: number) => void;
-  detectEnvironment: () => Record<string, boolean>;
 
   package_loader: any;
   importlib: any;
@@ -593,7 +585,7 @@ export type PackageManagerAPI = Pick<
  * @hidden
  */
 export type PackageManagerModule = Pick<
-  Module,
+  PyodideModule,
   | "PATH"
   | "LDSO"
   | "stringToNewUTF8"